feat: 分割したプレイ画面ロジックとUI
