+++ components/ResultModal.tsx
@@ -1,0 +1,70 @@
+import React from 'react';
+import { Modal, StyleSheet, View } from 'react-native';
+import { ThemedText } from '@/components/ThemedText';
+import { ThemedView } from '@/components/ThemedView';
+import { PlainButton } from '@/components/PlainButton';
+import type { HighScore } from '@/src/game/highScore';
+
+/** 結果表示モーダル */
+export function ResultModal({
+  visible,
+  top,
+  title,
+  steps,
+  bumps,
+  stageText,
+  highScore,
+  newRecord,
+  onOk,
+  okLabel,
+  accLabel,
+}: {
+  visible: boolean;
+  top: number;
+  title: string;
+  steps: string;
+  bumps: string;
+  stageText: string;
+  highScore: HighScore | null;
+  newRecord: boolean;
+  onOk: () => void | Promise<void>;
+  okLabel: string;
+  accLabel: string;
+}) {
+  return (
+    <Modal transparent visible={visible} animationType="fade">
+      <View style={styles.wrapper} accessible accessibilityLabel="結果表示オーバーレイ">
+        <ThemedView style={[styles.content, { marginTop: top }]}>
+          <ThemedText type="title">{title}</ThemedText>
+          <ThemedText>{steps}</ThemedText>
+          <ThemedText>{bumps}</ThemedText>
+          <ThemedText>{stageText}</ThemedText>
+          {highScore && (
+            <ThemedText>
+              Best: {highScore.stage}ステージ / {highScore.steps} steps / {highScore.bumps} bumps
+            </ThemedText>
+          )}
+          {newRecord && <ThemedText>記録更新！</ThemedText>}
+          <PlainButton title={okLabel} onPress={onOk} accessibilityLabel={accLabel} />
+        </ThemedView>
+      </View>
+    </Modal>
+  );
+}
+
+const styles = StyleSheet.create({
+  wrapper: {
+    flex: 1,
+    justifyContent: 'flex-start',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0,0,0,0.5)',
+  },
+  content: {
+    backgroundColor: '#fff',
+    padding: 20,
+    borderRadius: 8,
+    alignItems: 'center',
+    gap: 10,
+    width: 250,
+  },
+});
